pyproject.toml
[project]
name = "cdk-pupper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "aws-cdk-lib==2.201.0",
    "boto3>=1.39.1",
    "cdk-nag>=2.36.31",
    "constructs>=10.0.0,<11.0.0",
    "fastapi>=0.115.14",
    "mangum>=0.19.0",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "python-dateutil>=2.8.2",
    "cryptography>=45.0.5",
    "pydantic>=1.10.22",
    "typing-extensions>=4.14.0",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
